PsychOpenXR[Core]: Deal with unsupported OpenXR swapchain image formats.
A given OpenXR implementation may not support the requested OpenGL
swapchain texture internal format, so the driver has to fallback to
a different format. Return the actual chosen format to PsychOpenXR,
so it can pass it along to Screen's imaging pipeline. This way the
pipeline can adapt to a mismatch between it wanted - and was configured
for during 'OpenWindow' - and what it got as finalizedFBO's. E.g., if
the new finalizedFBO format is unsuitable to represent the drawBufferFBO
at requested range and precision, it may choose to unshare drawBufferFBO
and finalizedFBO and take the performance hit instead of messing up the
color precision and format.

A typical example would be requesting a float 16 drawbuffer, but the
OpenXR runtime can not do that, so one has to split into float 16
virtual framebuffer and other output buffer, as fp16 has different
range, precision and linearity behavior. As it turns out, while Linux
MonadoXR can do float16, Windows OculusXR runtime can't.

When we are at it, enumerate all supported swapchain formats in
descending priority order for debugging at verbosity 5+.